+++ tests/schema/extensions/test_validation_cache.py
